Fix access key API version 0.16.0
The response needs to be the response that was current at version 0.16.0
